Stacked: Fix incorrect CLK pin assignment
This pin is meant to be used to disable the external clock on IMU modules, it should be connected to CLK_CTL on Meia's modules.
This changes it to the pin adjacent to the IMU module pin in the stacked configuration

Confusingly, modules have a pin labeled CLK, but this is for outputting the actual clock signal.