Add stacked configurations
Adds device tree overlays to reconfigure the pins and define the GPIO power pins.

stacked.overlay is the base for configuring the power and button
stacked-i2c.overlay configures the i2c pins and disables spi
stacked-spi.overlay configures the spi pins and disables i2c

Apply both the stacked overlay and either stacked-spi or stacked-i2c depending on which protocol to use. No changes if the overlays are not used.